Defaul avatar, required fields, new tests

+++ routes/blogs.js
@@ -22,40 +22,54 @@ router.get('/', function (req, res) {
 
 //create new blog
 router.post('/', middleware.isLoggedIn, function (req, res) {
-    //upload image to cloud
-    cloudinary.uploader.upload(req.file.path, function (results) {
-        //delete file from tmp folder
-        fs.unlink(req.file.path);
-        //add to blog obj image cloud link
-        req.body.blog.image = results;
-        //scripts is not allowed in blog body
-        req.body.blog.body = req.sanitize(req.body.blog.body);
-        //looking for user in db
-        User.findById(req.user.id, function (err, foundUser) {
-            if (err) {
-                console.log(err);
-            } else {
-                //save blog in db
-                Blog.create(req.body.blog, function (err, newBlog) {
-                    if (err) {
-                        console.log(err)
-                    } else {
-                        //update blog author obj
-                        newBlog.author.id = req.user._id;
-                        newBlog.author.fname = req.user.fname;
-                        newBlog.author.lname = req.user.lname;
-                        newBlog.author.image = req.user.image;
-                        newBlog.save();
+    //scripts is not allowed in blog body
+    req.body.blog.body = req.sanitize(req.body.blog.body);
+    //looking for user in db
+    User.findById(req.user.id, function (err, foundUser) {
+        if (err) {
+            console.log(err);
+        } else {
+            //save blog in db
+            Blog.create(req.body.blog, function (err, newBlog) {
+                if (err) {
+                    console.log(err)
+                } else {
+                    async.series([
+                        function (callback) {
+                            if (req.file) {
+                                //upload image to cloud
+                                cloudinary.uploader.upload(req.file.path, function (results) {
+                                    //delete file from tmp folder
+                                    fs.unlink(req.file.path);
+                                    //add to blog obj image cloud link
+                                    req.body.blog.image = results;
+                                    callback();
+                                }, {type: "private", folder: "blogs/"});
+                            } else {
+                                callback();
+                            }
+                        },
+                        function (callback) {
+                            //update blog author obj
+                            newBlog.author.id = req.user._id;
+                            newBlog.author.fname = req.user.fname;
+                            newBlog.author.lname = req.user.lname;
+                            newBlog.author.image = req.user.image;
+                            //update blog image
+                            newBlog.image = req.body.blog.image;
+                            newBlog.save();
 
-                        //update users blogs array
-                        foundUser.blogs.push(newBlog);
-                        foundUser.save();
-                        res.redirect('/blogs/' + newBlog._id);
-                    }
-                })
-            }
-        })
-    }, {type: "private", folder: "blogs/"});
+                            //update users blogs array
+                            foundUser.blogs.push(newBlog);
+                            foundUser.save();
+                            res.redirect('/blogs/' + newBlog._id);
+                            callback();
+                        }
+                    ]);
+                }
+            })
+        }
+    })
 });
 //return page to create new blog
 router.get('/new', middleware.isLoggedIn, function (req, res) {
@@ -116,46 +130,54 @@ router.delete('/:id', middleware.checkBlogOwnership, function (req, res) {
         if (err) {
             console.log(err);
         } else {
-            cloudinary.uploader.destroy(blog.image.public_id, function (result) {
-                async.series([
-                    function (callback) {
-                        //async remove for comments array
-                        async.each(blog.comments, function (comment, callback) {
-                            Comment.findByIdAndRemove(comment, function (err) {
-                                if (err) {
-                                    console.log(err);
-                                    callback('Cant find a comment')
-                                } else {
-                                    callback();
-                                }
-                            })
-                        }, function (err) {
+            async.series([
+                //if blog has image it will be removed from cloud
+                function (callback) {
+                    if (blog.image) {
+                        cloudinary.uploader.destroy(blog.image.public_id, function (result) {
+                            callback();
+                        }, {type: "private"});
+                    } else {
+                        callback();
+                    }
+                },
+                function (callback) {
+                    //async remove for comments array
+                    async.each(blog.comments, function (comment, callback) {
+                        Comment.findByIdAndRemove(comment, function (err) {
                             if (err) {
-                                console.log('A file failed to process!');
+                                console.log(err);
+                                callback('Cant find a comment')
                             } else {
                                 callback();
                             }
                         })
-                    },
-                    //Remove blogId from foundUser blogs arr
-                    function (callback) {
-                        User.findById(req.user.id, function (err, foundUser) {
-                            var index = foundUser.blogs.indexOf(blog._id);
-                            if (index > -1) {
-                                foundUser.blogs.splice(index, 1);
-                                foundUser.save();
-                            } else {
-                                console.log('No such element');
-                            }
+                    }, function (err) {
+                        if (err) {
+                            console.log('A file failed to process!');
+                        } else {
                             callback();
-                        })
-                    },
-                    function (callback) {
-                        res.redirect('/blogs');
+                        }
+                    })
+                },
+                //Remove blogId from foundUser blogs arr
+                function (callback) {
+                    User.findById(req.user.id, function (err, foundUser) {
+                        var index = foundUser.blogs.indexOf(blog._id);
+                        if (index > -1) {
+                            foundUser.blogs.splice(index, 1);
+                            foundUser.save();
+                        } else {
+                            console.log('No such element');
+                        }
                         callback();
-                    }
-                ]);
-            }, {type: "private"});
+                    })
+                },
+                function (callback) {
+                    res.redirect('/blogs');
+                    callback();
+                }
+            ]);
         }
     })
 });
